competitive_intel/requirements.txt
anthropic>=0.40.0
notion-client>=2.2.1
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.4
tenacity>=8.2.3
